Allow repositories to fetch their builds
In order to achieve that, we had to change how our API was working.

If we want to fetch a single build, we would have to call that route:
"builds/{ID}". Then, if we wanted to fetch all the builds associated
with the repository we would have to call: "repos/{repo_slug}/builds".

If we were to call:

`Travis::Build.find_all("#{repo_slug}/builds")` the actual API call that
would have been made would be "builds/{repo_slug}/builds"
which doesn't make any sense from the point of view of Travis.

That is due to the fact that we are prepending or appending the resource
automatically when calling the finder.

Since we want to use the same API convention, I've opted to add a new
parameters to the finder methods which is named option and we can pass
if we want to automatically include the resource name when calling the
API. In the case of the repository fetching their builds, we don't want
to append the "build" ressource_name at the beginning of the request
because the API call being used to fetch the build from a ressource is
different from when we fetch builds standalone.

This commit also includes some refactoring.